Most GUI and Model functionality implemented now.
View: observers are in place to update in response to changes in the
model (SlotLabel and StatusLabel observe Spreadsheet).

Controller: listeners are in place to react to user events (Editor,
StatusLabel, and Spreadsheet listen for events and exceptions)

Model: The strategy to prevent self-reference and other issues is
implemented as a method in Spreadsheet. Spreadsheet doesn't have a clear
method, but with the conditional-change method in place, implementation
should be fairly trivial. The insert() method in Spreadsheet has been
cleaned up dramatically with help from a SlotFactory class. It will
create the correct Slot for the correct String (i.e. comment ->
CommentSlot, expression -> ExpressionSlot, emptyString -> DummySlot).

GUI.menu: Completely untouched... Should double check that...